some sed-issues

+++ triple_distilled.pl
@@ -23,26 +23,46 @@ foreach my $QUERY (@QUERY){
         chdir($values[3]);
        
 # CELERA-mapping
-       # get fastq of all reads
+
+    # get fastq of all reads
+    if (-e "all\.fastq") {
+        print "File all\.fastq exists.\n";
+    }
+    else{
+
        `bedtools bamtofastq \-i $values[0] \-fq alll\.fastq`; 
        `sed  \'\/\^\@$values[1]\/ s\/\[\:\/\]\[12\]\$\/\/\' alll\.fastq \> all\.fastq`; 
        `rm alll\.fastq`; 
+
+    }       
       # get headers frmo the remaining unmapped reads
+    if (-e "name\_unmapped\.lst") {
+        print "File name\_unmapped\.lst exists.\n";
+    }
+    else{
        `sed  \'\/\^\@$values[1]\/ s\/\[\:\/\]\[12\]\$\/\/\' \/data\_fedor12\/robin\/Q\_C\_Y\_C\/$values[3]\/1\_\* \| grep \-o \"\@$values[1]\[\^\ \  \]\*\" \| sort \| uniq \| sed \'s\/\^\@\/\/\' \| sed \'s\/\[\:\/\]\[12\]\$\/\/\' \> name\_unmapped\.lst`;
+    }  
        # get interleaved fastq of unmapped reads
-       `seqtk subseq all\.fastq name\_unmapped\.lst \> ready\_for\_celera\_mapping\.fastq`;
+    if (-e "ready\_for\_celera\_mapping\.fastq") {
+        print "File ready\_for\_celera\_mapping\.fastq exists.\n";
+    }
+    else{
+        `seqtk subseq all\.fastq name\_unmapped\.lst \> ready\_for\_celera\_mapping\.fastq`;
+    }  
+      
+      
         # map to celera in paired-end mode
         `\/home\/robin\/bin\/bwa\-0\.7\.5a\/bwa mem \-R \'\@RG\\\tID\:$values[3]\\tSM\:$values[3]\' \-M \-p \/data\_fedor12\/robin\/databases\/Celera\/Alt\_Rn\_Celera\.fa ready\_for\_celera\_mapping\.fastq \> Celera\.sam`;
         # get reads, that properly map in pairs
         `samtools view \-bS \-f 2 Celera\.sam \> Celera\_proper\_mapped\.bam`;
         # get headers from properly mapped readpairs
-        `samtools view \-f 2  Celera\_proper\_mapped\.bam \| awk \'\{print \$1\}\' \| sort \| uniq \> nameM\.lst`;
+        `samtools view \-f 2  Celera\_proper\_mapped\.bam \| awk \'\{print \$1\}\' \| sort \| uniq \> nameMc\.lst`;
         #get headers from not properly mapped readpairs
-        `samtools view \-F 2  Celera\_proper\_mapped\.bam \| awk \'\{print \$1\}\' \| sort \| uniq \> nameU\.lst`;
+        `samtools view \-F 2  Celera\_proper\_mapped\.bam \| awk \'\{print \$1\}\' \| sort \| uniq \> nameUc\.lst`;
         #get fastq of properly mapped readpairs
-        `seqtk subseq all\.fastq nameM\.lst \> C\.fastq`;
+        `seqtk subseq all\.fastq nameMc\.lst \> C\.fastq`;
         #get fastq of properly mapped readpairs
-        `seqtk subseq all\.fastq nameU\.lst \> Cu\.fastq`;
+        `seqtk subseq all\.fastq nameUc\.lst \> Cu\.fastq`;
         # stats of mapped readpairs
         `echo \'\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#     Celera\-mapped readpairs     \#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\' \>\> TripleDistilled\.log`;
         `\/home\/robin\/bin\/fasta\_utilities\/ea\-utils\.\1\.\1\.2\-537\/fastq\-stats C\.fastq \>\> TripleDistilled\.log`;
@@ -61,13 +81,13 @@ foreach my $QUERY (@QUERY){
         # get reads, that properly map in pairs
         `samtools view \-bS \-f 2 Ychr\.sam \> Ychr\_proper\_mapped\.bam`;
         # get headers from properly mapped readpairs
-        `samtools view \-f 2  Ychr\_proper\_mapped\.bam \| awk \'\{print \$1\}\' \| sort \| uniq \> nameM\.lst`;
+        `samtools view \-f 2  Ychr\_proper\_mapped\.bam \| awk \'\{print \$1\}\' \| sort \| uniq \> nameMy\.lst`;
         #get headers from not properly mapped readpairs
-        `samtools view \-F 2  Ychr\_proper\_mapped\.bam \| awk \'\{print \$1\}\' \| sort \| uniq \> nameU\.lst`;
+        `samtools view \-F 2  Ychr\_proper\_mapped\.bam \| awk \'\{print \$1\}\' \| sort \| uniq \> nameUy\.lst`;
         #get fastq of properly mapped readpairs
-        `seqtk subseq all\.fastq nameM\.lst \> Y\.fastq`;
+        `seqtk subseq all\.fastq nameMy\.lst \> Y\.fastq`;
         #get fastq of properly mapped readpairs
-        `seqtk subseq all\.fastq nameU\.lst \> Yu\.fastq`;
+        `seqtk subseq all\.fastq nameUy\.lst \> Yu\.fastq`;
         # stats of mapped readpairs
         `echo \'\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#     Ychr\-mapped readpairs     \#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\' \>\> TripleDistilled\.log`;
         `\/home\/robin\/bin\/fasta\_utilities\/ea\-utils\.\1\.\1\.2\-537\/fastq\-stats Y\.fastq \>\> TripleDistilled\.log`;
@@ -88,13 +108,13 @@ foreach my $QUERY (@QUERY){
         # extract properly-mapped reads
         `samtools view \-bS \-f 2 V\.sam \> V\_proper\_mapped\.bam`;
         # get headers from properly mapped readpairs
-        `samtools view \-f 2  V\_proper\_mapped\.bam \| awk \'\{print \$1\}\' \| sort \| uniq \> nameM\.lst`;
+        `samtools view \-f 2  V\_proper\_mapped\.bam \| awk \'\{print \$1\}\' \| sort \| uniq \> nameMv\.lst`;
         #get headers from not properly mapped readpairs
-        `samtools view \-F 2 V\_proper\_mapped\.bam \| awk \'\{print \$1\}\' \| sort \| uniq \> nameU\.lst`;
+        `samtools view \-F 2 V\_proper\_mapped\.bam \| awk \'\{print \$1\}\' \| sort \| uniq \> nameUv\.lst`;
         #get fastq of properly mapped readpairs
-        `seqtk subseq all\.fastq nameM\.lst \> V\.fastq`;
+        `seqtk subseq all\.fastq nameMv\.lst \> V\.fastq`;
         #get fastq of properly mapped readpairs
-        `seqtk subseq all\.fastq nameU\.lst \> Vu\.fastq`;
+        `seqtk subseq all\.fastq nameUv\.lst \> Vu\.fastq`;
         # stats of mapped readpairs
         `echo \'\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#     ViPro\-mapped readpairs     \#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\#\' \>\> TripleDistilled\.log`;
         `\/home\/robin\/bin\/fasta\_utilities\/ea\-utils\.\1\.\1\.2\-537\/fastq\-stats V\.fastq \>\> TripleDistilled\.log`;
